Move methods that deal with the current user to the CurrentUserController.
